requirements.txt
# Github
GitPython

# Core LangChain libraries
langchain
langchain-core
langchain-community
langchain-chroma

# OpenAI integration for embeddings
langchain-openai
openai

